requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langgraph>=0.3.0,<1.1.0
streamlit
requests

# embeddings stack
sentence-transformers>=2.2.0

chromadb>=0.4.22
pypdf>=3.17.0

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.2

pandas>=2.1.0
numpy>=1.24.0
